AALP-1699 Updating Call log entity and service changes for storing Voice files
Changes done -  to include 3 new fields; external id, external type - to store any of the service provider maintained ids, and playedMessages, which can be used to store all the voice files played in each of the node. Related service changes to store the values in these 3 fields.

Note : This would store all the messages to be played at the node level. i.e., when the update happens, at a shot all the audio files present under that node would be saved.